test: fix unit tests, skip exposure tests for now
