+++ tests/interactions/sex_test.py
@@ -18,7 +18,7 @@ def test_sex_transmission(make_model, make_agent):
     p.partners["Sex"] = set()
     rel = Relationship(a, p, 10, bond_type="Sex")
 
-    a.hiv = True
+    a.hiv.active = True
     a.hiv.time = model.time  # acute
 
     rel.total_sex_acts = 0
@@ -30,22 +30,21 @@ def test_sex_transmission(make_model, make_agent):
 
     # test partner becomes
     Sex.interact(model, rel)
-    assert p.hiv
+    assert p.hiv.active
 
 
 @pytest.mark.unit
-def test_sex_transmission_do_nothing(make_model, make_agent):
+def test_sex_num_acts(make_model, make_agent, make_relationship, params):
+    params.hiv.dx.risk_reduction.sex = 1.0
     model = make_model()
+    model.np_random = FakeRandom(1.0)
     a = make_agent()
     p = make_agent()
-    a.partners["Sex"] = set()
-    p.partners["Sex"] = set()
-    rel = Relationship(a, p, 10, bond_type="Sex")
+    rel = make_relationship(a, p)
 
-    assert Sex.interact(model, rel) is False
+    assert Sex.get_num_acts(model, rel) > 0
 
-    a.hiv = True
-    p.hiv = True
+    a.hiv.active = True
+    a.hiv.dx = True
 
-    # test nothing happens
-    assert Sex.interact(model, rel) is False
+    assert Sex.get_num_acts(model, rel) == 0
