+++ tests/model_test.py
@@ -45,7 +45,7 @@ def test_update_AllAgents(make_model, make_agent):
     model.time = 1
     # update all agents passes with agent 0
     model.update_all_agents()
-    assert a.pca.awareness is False
+    assert a.knowledge.active is False
 
     # remove all bonds compatible with agent 0. agent 0 fails
     for rel in copy(model.pop.relationships):
@@ -54,210 +54,6 @@ def test_update_AllAgents(make_model, make_agent):
     with pytest.raises(ValueError) as excinfo:
         model.update_all_agents()
     assert "No agent zero!" in str(excinfo)
-
-
-@pytest.mark.unit
-def test_agents_interact(make_model, make_agent):
-    model = make_model()
-    a = make_agent(race="white", SO="HM")
-    p = make_agent(race="white", SO="HF")
-    a.partners["Sex"] = set()
-    p.partners["Sex"] = set()
-    rel = Relationship(a, p, 10, bond_type="Sex")
-
-    model.run_random = FakeRandom(0.6)
-    model.time = 0
-
-    a.incar.active = True
-    assert model.agents_interact(rel) is False
-
-    a.incar.active = False
-    assert model.agents_interact(rel) is False  # neither HIV
-
-    model.hiv_convert(a)
-    model.hiv_convert(p)
-    assert model.agents_interact(rel) is False  # both HIV
-
-    p.hiv = False
-
-    assert model.agents_interact(rel)  # sex transmssion
-    assert p.hiv is False  # but nothing happened (see skipped test)
-
-    a.drug_type = "Inj"
-    p.drug_type = "Inj"
-    rel.bond_type = "Inj"
-
-    model.run_random = FakeRandom(-0.1)
-
-    assert model.agents_interact(rel)  # needle transmission
-    assert p.hiv
-
-    p.hiv = False
-    model.run_random = FakeRandom(1.1)
-
-    assert model.agents_interact(rel)  # needle and sex
-    assert p.hiv is False  # but nothing happened
-
-
-@pytest.mark.unit
-def test_get_transmission_probability(make_model, make_agent):
-    model = make_model()
-    a = make_agent(race="white", SO="MSM")
-    a.haart.adherence = 1
-    a.sex_role = "versatile"
-
-    p = make_agent(race="white", SO="MSM")
-    p.sex_role = "versatile"
-    p.haart.adherence = 1
-
-    # test versatile-versatile relationship
-    p_needle = (
-        model.params.partnership.injection.transmission.haart_scaling[1].scale
-        * model.params.partnership.injection.transmission.base
-    )
-    p_sex = (
-        model.params.partnership.sex.haart_scaling["MSM"][1].prob
-        * model.params.partnership.sex.acquisition.MSM.versatile
-    )
-    scale = model.params.calibration.acquisition
-
-    assert model.get_transmission_probability("injection", a, p) == p_needle * scale
-    assert model.get_transmission_probability("sex", a, p) == p_sex * scale
-
-    # test one vers agent, one receptive agent
-    a.sex_role = "receptive"
-    p_sex_ins = (
-        model.params.partnership.sex.haart_scaling.MSM[1].prob
-        * model.params.partnership.sex.acquisition.MSM.insertive
-    )
-    p_sex_rec = (
-        model.params.partnership.sex.haart_scaling.MSM[1].prob
-        * model.params.partnership.sex.acquisition.MSM.receptive
-    )
-
-    assert model.get_transmission_probability("sex", a, p) == p_sex_ins * scale
-    assert model.get_transmission_probability("sex", p, a) == p_sex_rec * scale
-
-
-@pytest.mark.unit
-def test_hiv_convert(make_model, make_agent):
-    model = make_model()
-    a = make_agent()
-    a.prep.active = True
-
-    model.run_random = FakeRandom(-0.1)
-
-    model.hiv_convert(a)
-
-    assert a.hiv
-    assert a.hiv.time == model.time
-    assert a in model.pop.hiv_agents.members
-    assert a.prep.active is False
-
-
-@pytest.mark.unit
-def test_diagnose_hiv(make_model, make_agent):
-    model = make_model()
-    model.params.partner_tracing.prob = 1.0
-    model.time = 1
-    a = make_agent()
-    p = make_agent()
-    p.hiv = True
-    a.partners["Sex"].add(p)
-
-    model.run_random = FakeRandom(1.1)  # always greater than param
-    model.diagnose_hiv(a)
-
-    assert a.hiv.dx is False
-    assert p.hiv.dx is False
-    assert not p.partner_traced
-
-    model.run_random = FakeRandom(-0.1)  # always less than param
-    model.diagnose_hiv(a)
-
-    assert a.hiv.dx
-    assert a.hiv.dx_time == model.time
-    assert p in a.get_partners()
-    assert p.partner_traced
-    assert p.trace_time == model.time
-
-    assert p.hiv.dx is False
-    model.params.demographics[p.race][p.sex_type].hiv.dx.prob = 0
-
-    model.time = p.partner_traced + 1
-    model.diagnose_hiv(p)
-    assert p.hiv.dx
-    assert p.partner_traced is False
-
-
-@pytest.mark.unit
-def test_diagnose_hiv_already_tested(make_model, make_agent):
-    model = make_model()
-    a = make_agent()
-
-    a.hiv.dx = True
-
-    model.run_random = FakeRandom(-0.1)  # always less than param
-    model.diagnose_hiv(a)
-
-    assert a.hiv.dx
-
-
-@pytest.mark.unit
-def test_progress_to_aids_error(make_agent, make_model):
-    a = make_agent()
-    a.hiv = False
-    model = make_model()
-    a.target_partners = {bond: 0 for bond in model.params.classes.bond_types.keys()}
-    model.pop.add_agent(a)
-    num_aids = sum([1 for agent in model.pop.hiv_agents if agent.aids])  # get baseline
-
-    # test error case, agent must be HIV+
-    with pytest.raises(AssertionError):
-        model.progress_to_aids(a)
-
-    assert sum([1 for agent in model.pop.hiv_agents if agent.aids]) == num_aids
-
-
-@pytest.mark.unit
-def test_progress_to_aids_nothing(make_agent, make_model):
-    a = make_agent()
-    a.hiv = True
-    model = make_model()
-    a.target_partners = {bond: 0 for bond in model.params.classes.bond_types.keys()}
-    model.pop.add_agent(a)
-    num_aids = sum([1 for agent in model.pop.hiv_agents if agent.aids])  # get baseline
-
-    # test nothing case
-    a.hiv = True
-    a.haart.adherence = 1  # .0051 prob
-
-    model.run_random = FakeRandom(0.9)  # no AIDS
-
-    assert model.progress_to_aids(a) is None
-    assert sum([1 for agent in model.pop.hiv_agents if agent.aids]) == num_aids
-    assert a.aids is False
-
-
-@pytest.mark.unit
-def test_progress_to_aids_progress(make_agent, make_model):
-    a = make_agent()
-    a.hiv = True
-    model = make_model()
-    a.target_partners = {bond: 0 for bond in model.params.classes.bond_types.keys()}
-    model.pop.add_agent(a)
-    num_aids = sum([1 for agent in model.pop.hiv_agents if agent.aids])  # get baseline
-    a.location.params.hiv.aids.prob = 1.0
-
-    a.hiv = True
-    a.haart.adherence = 1  # .0051 prob
-
-    # test progress case
-    model.run_random = FakeRandom(0.001)  # AIDS
-
-    assert model.progress_to_aids(a) is None
-    assert sum([1 for agent in model.pop.hiv_agents if agent.aids]) == num_aids + 1
-    assert a.aids is True
 
 
 @pytest.mark.unit
