+++ tests/population_io_test.py
@@ -6,15 +6,17 @@ from copy import deepcopy
 from titan.population_io import (
     write,
     read,
-    agent_core_attrs,
     agent_exclude_attrs,
+    agent_feature_attrs,
+    agent_exposure_attrs,
     find_agent,
 )
 from titan.features import Prep, BaseFeature
+from titan.exposures import BaseExposure
 
 
 @pytest.mark.unit
-def test_write_pop_core(tmpdir, make_population):
+def test_write_pop(tmpdir, make_population):
     pop = make_population(n=10)
 
     write(pop, tmpdir, compress=False)
@@ -33,44 +35,15 @@ def test_write_pop_core(tmpdir, make_population):
             res.append(row)
 
     for row, a in zip(res, pop.all_agents):
-        for attr in agent_core_attrs:
-            assert repr(getattr(a, attr)) == row[attr]
-
-    # relationship writing
-    res = []
-    with open(rel_file, newline="") as f:
-        reader = csv.DictReader(f)
-        for row in reader:
-            res.append(row)
-
-    assert len(res) > 0
-    for row, r in zip(res, pop.relationships):
-        for attr in r.__dict__.keys():
-            assert repr(getattr(r, attr)) == row[attr]
-
-
-@pytest.mark.unit
-def test_write_pop_intervention(tmpdir, make_population):
-    pop = make_population(n=10)
-
-    write(pop, tmpdir, intervention_attrs=True, compress=False)
-
-    agent_file = os.path.join(tmpdir, f"{pop.id}_agents.csv")
-    assert os.path.isfile(agent_file)
-
-    rel_file = os.path.join(tmpdir, f"{pop.id}_relationships.csv")
-    assert os.path.isfile(rel_file)
-
-    # agents writing
-    res = []
-    with open(agent_file, newline="") as f:
-        reader = csv.DictReader(f)
-        for row in reader:
-            res.append(row)
-
-    for row, a in zip(res, pop.all_agents):
         for attr in a.__dict__.keys():
-            if attr in agent_exclude_attrs:
+            if any(
+                attr in exclude_set
+                for exclude_set in (
+                    agent_exclude_attrs,
+                    agent_exposure_attrs,
+                    agent_feature_attrs,
+                )
+            ):
                 assert attr not in row
             else:
                 assert repr(getattr(a, attr)) == row[attr]
@@ -88,33 +61,11 @@ def test_write_pop_intervention(tmpdir, make_population):
 
 
 @pytest.mark.unit
-def test_read_pop_core(tmpdir, make_population, params):
-    pop = make_population(n=10)
-
-    write(pop, tmpdir, compress=False)
-
-    new_pop = read(params, tmpdir)
-
-    assert pop.id == new_pop.id
-    assert pop.all_agents.num_members() == new_pop.all_agents.num_members()
-    assert len(pop.relationships) == len(new_pop.relationships)
-
-    agent = next(iter(pop.all_agents))
-    new_agent = find_agent(new_pop, str(agent.id))
-
-    for attr in agent_core_attrs:
-        assert getattr(agent, attr) == getattr(new_agent, attr)
-
-    assert agent.relationships == new_agent.relationships
-    assert agent.partners == new_agent.partners
-
-
-@pytest.mark.unit
-def test_read_pop_intervention(tmpdir, make_population, params):
+def test_read_pop(tmpdir, make_population, params):
     pop = make_population(n=10)
     prep_counts = deepcopy(Prep.counts)
 
-    write(pop, tmpdir, intervention_attrs=True, compress=False)
+    write(pop, tmpdir, compress=False)
 
     new_pop = read(params, tmpdir)
 
@@ -135,6 +86,10 @@ def test_read_pop_intervention(tmpdir, make_population, params):
             feat_attrs = orig_attr.__dict__.keys()
             for feat_attr in feat_attrs:
                 assert getattr(orig_attr, feat_attr) == getattr(new_attr, feat_attr)
+        elif isinstance(orig_attr, BaseExposure):
+            expose_attrs = orig_attr.__dict__.keys()
+            for expose_attr in expose_attrs:
+                assert getattr(orig_attr, expose_attr) == getattr(new_attr, expose_attr)
         else:
             assert orig_attr == new_attr
 
@@ -145,7 +100,7 @@ def test_write_read_pop_compressed(tmpdir, make_population, params):
     pop = make_population(n=10)
     prep_counts = deepcopy(Prep.counts)
 
-    archive = write(pop, tmpdir, intervention_attrs=True, compress=True)
+    archive = write(pop, tmpdir, compress=True)
 
     new_pop = read(params, archive)
 
@@ -166,5 +121,9 @@ def test_write_read_pop_compressed(tmpdir, make_population, params):
             feat_attrs = orig_attr.__dict__.keys()
             for feat_attr in feat_attrs:
                 assert getattr(orig_attr, feat_attr) == getattr(new_attr, feat_attr)
+        elif isinstance(orig_attr, BaseExposure):
+            expose_attrs = orig_attr.__dict__.keys()
+            for expose_attr in expose_attrs:
+                assert getattr(orig_attr, expose_attr) == getattr(new_attr, expose_attr)
         else:
             assert orig_attr == new_attr
