+++ titan/features/vaccine.py
@@ -88,8 +88,8 @@ class Vaccine(base_feature.BaseFeature):
             time: the current model time step
             interaction_type: The type of interaction where the agent could acquire HIV (e.g. 'sex', 'injection' - from [params.classes.interaction_types])
         """
+        # not protected the time step the agent is vaccinaetd
         if self.active and self.time is not None and self.time < time:
-            # TO_REVIEW how to handle when time == self.time? skip?
             vaccine_time_months = (
                 (time - self.time)
                 / self.agent.location.params.model.time.steps_per_year
